Further refactorings: move classes to other packages
