+++ core/src/main/java/bisq/core/app/BisqEnvironment.java
@@ -17,9 +17,9 @@
 
 package bisq.core.app;
 
-import bisq.core.btc.BaseCurrencyNetwork;
 import bisq.core.btc.BtcOptionKeys;
-import bisq.core.btc.UserAgent;
+import bisq.core.btc.network.BaseCurrencyNetwork;
+import bisq.core.btc.network.UserAgent;
 import bisq.core.dao.DaoOptionKeys;
 import bisq.core.exceptions.BisqException;
 import bisq.core.filter.FilterManager;
