+++ core/src/main/java/bisq/core/btc/network/BitcoinNodes.java
@@ -15,7 +15,7 @@
  * along with bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.btc.setup;
+package bisq.core.btc.network;
 
 import bisq.core.app.BisqEnvironment;
