+++ core/src/main/java/bisq/core/btc/setup/BsqWallet.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.btc.wallet;
+package bisq.core.btc.setup;
 
 import org.bitcoinj.core.NetworkParameters;
 import org.bitcoinj.wallet.KeyChainGroup;
