+++ core/src/main/java/bisq/core/btc/wallet/RawTransactionInput.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.btc.data;
+package bisq.core.btc.wallet;
 
 import bisq.common.proto.network.NetworkPayload;
 import bisq.common.proto.persistable.PersistablePayload;
