+++ core/src/main/java/bisq/core/filter/FilterManager.java
@@ -19,7 +19,7 @@ package bisq.core.filter;
 
 import bisq.core.app.AppOptionKeys;
 import bisq.core.app.BisqEnvironment;
-import bisq.core.btc.setup.BitcoinNodes;
+import bisq.core.btc.network.BitcoinNodes;
 import bisq.core.payment.payload.PaymentAccountPayload;
 import bisq.core.payment.payload.PaymentMethod;
 import bisq.core.provider.ProvidersRepository;
