+++ core/src/main/java/bisq/core/trade/protocol/tasks/buyer_as_maker/BuyerAsMakerCreatesAndSignsDepositTx.java
@@ -17,10 +17,10 @@
 
 package bisq.core.trade.protocol.tasks.buyer_as_maker;
 
-import bisq.core.btc.data.PreparedDepositTxAndMakerInputs;
-import bisq.core.btc.data.RawTransactionInput;
 import bisq.core.btc.wallet.AddressEntry;
 import bisq.core.btc.wallet.BtcWalletService;
+import bisq.core.btc.wallet.PreparedDepositTxAndMakerInputs;
+import bisq.core.btc.wallet.RawTransactionInput;
 import bisq.core.offer.Offer;
 import bisq.core.trade.Trade;
 import bisq.core.trade.protocol.TradingPeer;
