+++ core/src/main/java/bisq/core/trade/protocol/tasks/buyer_as_taker/BuyerAsTakerCreatesDepositTxInputs.java
@@ -17,9 +17,9 @@
 
 package bisq.core.trade.protocol.tasks.buyer_as_taker;
 
-import bisq.core.btc.data.InputsAndChangeOutput;
 import bisq.core.btc.wallet.AddressEntry;
 import bisq.core.btc.wallet.BtcWalletService;
+import bisq.core.btc.wallet.InputsAndChangeOutput;
 import bisq.core.trade.Trade;
 import bisq.core.trade.protocol.tasks.TradeTask;
