+++ core/src/main/java/bisq/core/trade/protocol/tasks/seller_as_taker/SellerAsTakerSignAndPublishDepositTx.java
@@ -17,10 +17,10 @@
 
 package bisq.core.trade.protocol.tasks.seller_as_taker;
 
-import bisq.core.btc.data.RawTransactionInput;
 import bisq.core.btc.exceptions.TxBroadcastException;
 import bisq.core.btc.wallet.AddressEntry;
 import bisq.core.btc.wallet.BtcWalletService;
+import bisq.core.btc.wallet.RawTransactionInput;
 import bisq.core.btc.wallet.TxBroadcaster;
 import bisq.core.trade.Contract;
 import bisq.core.trade.Trade;
