+++ desktop/src/main/java/bisq/desktop/main/settings/network/NetworkSettingsView.java
@@ -29,7 +29,7 @@ import bisq.desktop.main.overlays.windows.TorNetworkSettingsWindow;
 import bisq.desktop.util.GUIUtil;
 
 import bisq.core.app.BisqEnvironment;
-import bisq.core.btc.setup.BitcoinNodes;
+import bisq.core.btc.network.BitcoinNodes;
 import bisq.core.btc.setup.WalletsSetup;
 import bisq.core.filter.Filter;
 import bisq.core.filter.FilterManager;
